requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
httpx
python-dotenv
base64